This is a snapshot in the middle of development of heapsort, not fit for use.
Makefile: added _GNU_SOURCE and _XOPEN_SOURCE=700
Added header and empty definitions for heapsort and mergesort.
Added testset for heapsort.

I replaced the heapsort implementation I have because it is not finished yet.
These changes are for a pull request to check on debug messages from bstring.
I still wonder why the test for heapset passes: it should fail as it doesn't sort.